Fix x64 integer width and pointer truncation throughout
quakeparms_t::memsize: int -> size_t; switch GlobalMemoryStatus to
GlobalMemoryStatusEx (DWORDLONG fields) to read physical RAM correctly
on machines with >2GB free memory.

zone.c: hunk tracking variables (hunk_size, hunk_low_used,
hunk_high_used, hunk_tempmark) -> size_t; propagate through zone.h
and every caller that stored a hunk mark.

sys_win.c: -heapsize multiply promoted to size_t to avoid int overflow;
QHOST handle args parsed with strtoull to preserve 64-bit handle values;
Sys_PageIn loop vars -> size_t; Sys_MakeCodeWriteable addr param ->
size_t (address must be pointer-wide).

cd_win.c: all mciSendCommand 4th-arg casts DWORD -> DWORD_PTR (winmm
derefs the pointer, truncating it to 32 bits caused the AV crash on
startup); dwCallback fields likewise.

gl_vidnt.c: MainWndProc return type LONG -> LRESULT (LRESULT is
pointer-sized on x64; LONG is always 32-bit).

net_main.c / net_vcr.c: VCR session field long -> intptr_t (field
stores a qsocket_t* for session identity; long is 32-bit on MSVC x64);
fix type-pun write in VCR_Connect.

common.c: alignment checks (long)ptr -> (size_t)ptr; long is 32-bit on
MSVC x64 so the high bits were lost, making the check accidentally work
only because alignment depends on low bits.

+++ WinQuake/cmd.c
@@ -283,7 +283,7 @@ Cmd_Exec_f
 void Cmd_Exec_f (void)
 {
 	char	*f;
-	int		mark;
+	size_t	mark;
 
 	if (Cmd_Argc () != 2)
 	{
